Add surveillance endpoint using live feed from cameras Implement websockets to identify client disconnect Use process sharing queue to put and get frames Stream live webcam by getting camera index ID Add a 5-minute timeout for robinhood endpoint Update .gitignore Use secrets.compare_digest to validate auth Remove display feature after capturing an image

+++ modules/telegram/bot.py
@@ -490,7 +490,7 @@ class TelegramBot:
             response: Response from Jarvis.
             payload: Payload received, to extract information from.
         """
-        if os.path.isfile(response):
+        if os.path.isfile(response) and response.endswith('jpg'):
             self.send_photo(chat_id=payload['from']['id'], filename=response)
             os.remove(response)
             return
